refactor: split selfupdate into build-tagged platform files (replace_unix.go, replace_windows.go)
