+++ internal/selfupdate/replace_test.go
@@ -4,22 +4,33 @@
 package selfupdate
 
 import (
-	"runtime"
+	"os"
+	"path/filepath"
 	"testing"
 )
 
-func TestPrepareSelfReplace_NonWindows(t *testing.T) {
-	if runtime.GOOS == "windows" {
-		t.Skip("this test verifies non-Windows no-op behavior")
+func TestCurrentExePath(t *testing.T) {
+	p, err := currentExePath()
+	if err != nil {
+		t.Fatalf("currentExePath() error: %v", err)
 	}
-	cleanup, err := PrepareSelfReplace()
+	if !filepath.IsAbs(p) {
+		t.Errorf("expected absolute path, got: %s", p)
+	}
+	if _, err := os.Stat(p); err != nil {
+		t.Errorf("resolved path should exist: %v", err)
+	}
+}
+
+func TestPrepareSelfReplace_ReturnsNoError(t *testing.T) {
+	// On any platform, calling PrepareSelfReplace should not panic.
+	restore, err := PrepareSelfReplace()
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
-	cleanup() // should be a no-op
+	restore()
 }
 
-func TestCleanupStaleFiles_NoError(t *testing.T) {
-	// Should not panic on any platform.
+func TestCleanupStaleFiles_NoPanic(t *testing.T) {
 	CleanupStaleFiles()
 }
